Refactor into multiple scripts
